+++ .eslintrc.js
@@ -3,11 +3,13 @@ require("@jobber/eslint-config/patch-eslint-plugin-resolution.js");
 
 const packageAliases = [
   ["@jobber/components", "./packages/components/src"],
+  ["@jobber/components-native", "./packages/components-native/src"],
   ["@jobber/hooks", "./packages/hooks/src"],
 ];
 
 module.exports = {
   extends: ["@jobber/eslint-config"],
+  root: true,
   settings: {
     "import/ignore": ["react-native/index"],
     "import/resolver": {
@@ -27,7 +29,14 @@ module.exports = {
     "import/no-internal-modules": [
       "error",
       {
-        allow: ["@jobber/components/*", "@jobber/hooks/*", "lodash/*"],
+        allow: [
+          "@jobber/components/*",
+          "@jobber/components-native",
+          "@jobber/hooks/*",
+          "@jobber/design/*",
+          "lodash/*",
+          "utils/*",
+        ],
       },
     ],
   },
@@ -36,11 +45,20 @@ module.exports = {
       files: ["*.stories.mdx"],
       extends: "plugin:mdx/recommended",
       rules: {
-        "react-native/no-inline-styles": "off",
         "no-alert": "off",
         "@typescript-eslint/naming-convention": "off",
         "@typescript-eslint/no-unused-expressions": "off",
         "import/no-extraneous-dependencies": "off",
+      },
+    },
+    {
+      files: ["*.stories.tsx"],
+      rules: {
+        "import/no-relative-parent-imports": "off",
+        "no-alert": "off",
+        "@typescript-eslint/naming-convention": "off",
+        "@typescript-eslint/no-unused-expressions": "off",
+        "import/no-default-export": "off",
       },
     },
   ],