+++ packages/components/src/FormField/FormFieldWrapper.tsx
@@ -54,6 +54,7 @@ export function FormFieldWrapper({
         // Naively assume that if the the type is tel, it is the InputPhoneNumber
         (placeholder && type === "tel"),
       [styles.textarea]: type === "textarea",
+      [styles.select]: type === "select",
       [styles.invalid]: invalid ?? error,
       [styles.disabled]: disabled,
       [styles.inline]: inline,