+++ packages/components/src/Grid/Grid.css.d.ts
@@ -1,0 +1,6 @@
+declare const styles: {
+  readonly "grid": string;
+  readonly "gap": string;
+};
+export = styles;
+