+++ packages/hooks/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.2](https://github.com/GetJobber/atlantis/compare/@jobber/hooks@2.1.1...@jobber/hooks@2.1.2) (2023-05-12)
+
+**Note:** Version bump only for package @jobber/hooks
+
+
+
+
+
 ## [2.1.1](https://github.com/GetJobber/atlantis/compare/@jobber/hooks@2.1.0...@jobber/hooks@2.1.1) (2023-05-01)
 
 **Note:** Version bump only for package @jobber/hooks